Fix flock implementation
Basically, the flock was failing to be acquired if there were
no file to lock it, this fix simply creates it if is not there
then tries to acquire it. There is a possible for race condition,
namely when to clients check together, and both find the file not
existet and try to create it, one of them fails, but this is too
small a window to worry about it (IMHO).